+++ src/services/progress-row/method-card.js
@@ -2,17 +2,11 @@
  * @module ProgressRow
  */
 
-import {
-  getDailySession,
-  getActivePath,
-  resetAllLearningPaths,
-  startLearningPath,
-  fetchLearningPathLessons,
-} from '../content-org/learning-paths'
+import { getActivePath, fetchLearningPathLessons } from '../content-org/learning-paths'
 import { getToday } from '../dateUtils.js'
-import { fetchByRailContentId, fetchByRailContentIds, fetchMethodV2IntroVideo } from '../sanity'
-import { addContextToContent } from '../contentAggregator.js'
+import { fetchMethodV2IntroVideo } from '../sanity'
 import { getProgressState } from '../contentProgress'
+import {COLLECTION_TYPE} from "../sync/models/ContentProgress.js";
 
 export async function getMethodCard(brand) {
   const introVideo = await fetchMethodV2IntroVideo(brand)
@@ -26,6 +20,10 @@ export async function getMethodCard(brand) {
   if (introVideoProgressState !== 'completed') {
     //startLearningPath('drumeo', 422533)
     const timestamp = Math.floor(Date.now() / 1000)
+    const instructorText =
+      introVideo.instructor?.length > 1
+        ? 'Multiple Instructors'
+        : introVideo.instructor?.[0]?.name || ''
     return {
       id: 1, // method card has no id
       type: 'method',
@@ -34,7 +32,7 @@ export async function getMethodCard(brand) {
       body: {
         thumbnail: introVideo.thumbnail,
         title: introVideo.title,
-        subtitle: `${introVideo.difficulty_string} • ${introVideo.instructor?.[0]?.name}`,
+        subtitle: `${introVideo.difficulty_string} • ${instructorText}`,
       },
       cta: {
         text: 'Get Started',
@@ -64,9 +62,7 @@ export async function getMethodCard(brand) {
     const nextIncompleteLesson = learningPath?.todays_lessons.find(
       (lesson) => lesson.progressStatus !== 'completed'
     )
-    let ctaText,
-      action,
-      nextLesson = null
+    let ctaText, action
     if (noneCompleted) {
       ctaText = 'Start Session'
       action = getMethodActionCTA(nextIncompleteLesson)
@@ -92,7 +88,7 @@ export async function getMethodCard(brand) {
 
     return {
       id: 1,
-      type: 'learning-path-v2',
+      type: COLLECTION_TYPE.LEARNING_PATH,
       progressType: 'method',
       header: 'Method',
       body: learningPath,